# api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/device/v1/device.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/addons/identify": {
      "post": {
        "summary": "openapi addons identify",
        "operationId": "Identify",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1AddonsIdentifyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddonsIdentifyRequest"
            }
          }
        ],
        "tags": [
          "Openapi"
        ]
      }
    },
...
    "/identify": {
      "get": {
        "summary": "openapi identify",
        "operationId": "Identify",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1IdentifyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Openapi"
        ]
      }
    },
    "/status": {
      "get": {
        "summary": "openapi status",
        "operationId": "Status",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1StatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Openapi"
        ]
      }
    },
...
    "/tenant/bind": {
      "post": {
        "summary": "openapi tenant bind",
        "operationId": "TenantBind",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1TenantBindResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1TenantBindRequst"
            }
          }
        ],
        "tags": [
          "Openapi"
        ]
      }
    },
    "/tenant/unbind": {
      "post": {
        "summary": "openapi tenant unbind",
        "operationId": "TenantUnbind",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1TenantUnbindResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1TenantUnbindRequst"
            }
          }
        ],
        "tags": [
          "Openapi"
        ]
      }
    }
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddonsIdentifyRequest": {
      "type": "object",
      "properties": {
        "plugin": {
          "$ref": "#/definitions/v1BriefPluginInfo"
        },
        "implemented_addons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ImplementedAddons"
          }
        }
      },
      "description": "*\r\nThe request value of the POST /v1/addons/identify,\r\nused when the plugin in the registration process implements\r\nthe extension point of the registered plugin."
    },
    "v1AddonsIdentifyResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        }
      },
      "description": "*\r\nThe response value of the POST /v1/addons/identify,\r\nused when the plugin in the registration process implements\r\nthe extension point of the registered plugin."
    },
    "v1AddonsPoint": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        }
      },
      "description": "*\r\nmessage plugin declares extension point."
    },
    "v1BriefPluginInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      },
      "description": "*\r\nmessage brief plugin information."
    },
...
    "v1IdentifyResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        },
        "plugin_id": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "tkeel_version": {
          "type": "string"
        },
        "addons_point": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1AddonsPoint"
          }
        },
        "implemented_plugin": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ImplementedPlugin"
          }
        }
      },
      "description": "*\r\nThe response value of the GET /v1/identify,\r\nused when the plugin registered."
    },
    "v1ImplementedAddons": {
      "type": "object",
      "properties": {
        "addons_point": {
          "type": "string"
        },
        "implemented_endpoint": {
          "type": "string"
        }
      },
      "description": "*\r\nmessage plugin declares implemented other plugin extension point."
    },
    "v1ImplementedPlugin": {
      "type": "object",
      "properties": {
        "plugin": {
          "$ref": "#/definitions/v1BriefPluginInfo"
        },
        "addons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ImplementedAddons"
          }
        }
      },
      "description": "*\r\nmessage plugin implemented other plugin."
    },
...
    "v1PluginStatus": {
      "type": "string",
      "enum": [
        "ERROR",
        "STARTING",
        "RUNING",
        "STOPPING",
        "STOPPED"
      ],
      "default": "ERROR",
      "description": "*\r\nplugin state enum."
    },
...
    "v1Result": {
      "type": "object",
      "properties": {
        "ret": {
          "$ref": "#/definitions/v1Retcode"
        },
        "msg": {
          "type": "string"
        }
      },
      "description": "*\r\nmessage result."
    },
    "v1Retcode": {
      "type": "string",
      "enum": [
        "OK",
        "BAD_REQEUST",
        "INTERNAL_ERROR"
      ],
      "default": "OK",
      "description": "*\r\nerror code enum."
    },
    "v1StatusResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        },
        "status": {
          "$ref": "#/definitions/v1PluginStatus"
        }
      },
      "description": "*\r\nThe request value of the GET /v1/Status.\r\nused when the tkeel checks the status of plugins regularly."
    },
...
    "v1TenantBindRequst": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string"
        },
        "extra": {
          "type": "string",
          "format": "byte"
        }
      },
      "description": "*\r\nThe request value of the POST /v1/tenant/bind,\r\nused when the tenant administrator enables the plugin."
    },
    "v1TenantBindResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        }
      },
      "description": "*\r\nThe response value of the POST /v1/tenant/bind,\r\nused when the tenant administrator enables the plugin."
    },
    "v1TenantUnbindRequst": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string"
        },
        "extra": {
          "type": "string",
          "format": "byte"
        }
      },
      "description": "*\r\nThe request value of the POST /v1/tenant/unbind,\r\nused when the tenant administrator disables the plugin."
    },
    "v1TenantUnbindResponse": {
      "type": "object",
      "properties": {
        "res": {
          "$ref": "#/definitions/v1Result"
        }
      },
      "description": "*\r\nThe response value of the POST /v1/tenant/unbind,\r\nused when the tenant administrator disables the plugin."
    },
...
  }
}